Refactored the socket reading code to used blocking sockets
git-svn-id: https://svn.apache.org/repos/asf/incubator/activemq/activemq-cpp/trunk/activemq-cpp@474558 13f79535-47bb-0310-9956-ffa450edef68

+++ src/test/activemq/io/BufferedOutputStreamTest.h
@@ -97,7 +97,7 @@ namespace io{
 		void testSmallerBuffer(){
 			
 			MyOutputStream myStream;
-			BufferedOutputStream bufStream( &myStream, 1 );
+			BufferedOutputStream bufStream( &myStream, (unsigned int)1 );
 			
 			const char* buffer = myStream.getBuffer();
 			
@@ -120,7 +120,7 @@ namespace io{
 		void testBiggerBuffer(){
 			
 			MyOutputStream myStream;
-			BufferedOutputStream bufStream( &myStream, 10 );
+			BufferedOutputStream bufStream( &myStream, (unsigned int)10 );
 			
 			const char* buffer = myStream.getBuffer();
